chore: rename annotations-folder and utilities
