+++ react-components/src/utilities/externalIndex.ts
@@ -3,4 +3,4 @@
  */
 export * from './instanceIds';
 export * from './instances';
-export * from './annotations';
+export * from './image360Annotations';
