+++ react-components/src/utilities/image360Annotations/equality.ts
