+++ react-components/src/utilities/image360Annotations/typeGuards.ts
